requirements.txt
asgiref==3.7.2
Django==5.0.2
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4